+++ R/as_sqlite_backend.R
@@ -12,28 +12,28 @@
 #'   See description.
 #' @param path (`NULL` | `character(1)`)\cr
 #'   Path for the SQLite data base. Defaults to a file in the temporary directory of the R session, see [tempfile()].
-#' @param ... (any)\cr
+#' @param ... (`any`)\cr
 #'   Additional arguments, currently ignored.
 #'
 #' @return [DataBackendDplyr].
 #' @export
-as_sqlite = function(data, path = NULL, ...) {
-  UseMethod("as_sqlite")
+as_sqlite_backend = function(data, path = NULL, ...) {
+  UseMethod("as_sqlite_backend")
 }
 
 #' @export
-as_sqlite.Task = function(data, path = NULL, ...) {
+as_sqlite_backend.Task = function(data, path = NULL, ...) {
   data$backend = sqlite_backend_from_data(cbind(data$data(), data.table(..row_id = data$row_ids)), path, "..row_id")
   data
 }
 
 #' @export
-as_sqlite.DataBackend = function(data, path = NULL, ...) {
+as_sqlite_backend.DataBackend = function(data, path = NULL, ...) {
   sqlite_backend_from_data(data$head(Inf), path, data$primary_key)
 }
 
 #' @export
-as_sqlite.data.frame = function(data, path = NULL, primary_key = "..row_id", ...) {
+as_sqlite_backend.data.frame = function(data, path = NULL, primary_key = "..row_id", ...) {
   assert_string(primary_key)
   if (primary_key %in% names(data)) {
     assert_atomic_vector(data[[primary_key]], unique = TRUE)
@@ -49,7 +49,9 @@ sqlite_backend_from_data = function(data, path, primary_key) {
   }
 
   assert_path_for_output(path)
-  on.exit( { if (file.exists(path)) file.remove(path) } )
+  on.exit({
+    if (file.exists(path)) file.remove(path)
+  })
 
   con = DBI::dbConnect(RSQLite::SQLite(), dbname = path, flags = RSQLite::SQLITE_RWC)
   DBI::dbWriteTable(con, "data", data)