+++ tests/testthat.R
@@ -1,4 +1,5 @@
-library(testthat)
-library(mlr3db)
-
-test_check("mlr3db")
+if (requireNamespace("testthat", quietly = TRUE)) {
+  library(testthat)
+  library(mlr3db)
+  test_check("mlr3db")
+}