+++ NEWS.md
@@ -1,3 +1,9 @@
-# mlr3db 0.1.0
+# mlr3db 0.1.2
+
+* Fixed `$distinct()` to not return missing values per default.
+* Added `na_rm` argument to `$distinct()`.
+* Renamed `as_sqlite()` to `as_sqlite_backend()`
+
+# mlr3db 0.1.1
 
 * Initial release.